Fix setup.py
Everything will be installed to gp_extras.kernels
