+++ README.md
@@ -20,7 +20,7 @@ You will need the new Gaussian process implementation from scikit-learn. For thi
 Now, you can install `kernel_extras`
 
     git clone git@github.com:jmetzen/gp_extras.git
-    cd gp_extras/gp_extras
+    cd gp_extras
     sudo python setup.py install
 
 Examples
